refactor: moved selectByPostType into qeta-common
